pass most arguments down when invoking across_all_types
In two places in datasets.py, when a method re-invoked itself via
across_all_types, it would drop the arguments passed down from the
caller.  In ProductPicker.make_product_list this was probably
harmless, but in CorpusFinalizer.create_and_upload_test_subset
it caused several of the command line arguments to “ix finalize”
to not be honored unless you also specified a product type.