Move notebooks out of pdr_tests package.
Also move mismatch_comparison.py and comparison_hook.py, which are
only used by one of the notebooks and which have no dependencies on
the rest of pdr_tests.  Also move the top level script
cover_everything_example.py into the notebooks directory.

The plan is to convert at least some of these notebooks and scripts to
proper ix subcommands in the future.  For now, moving them out of the
pdr_tests directory makes it clear that they are not part of the
package and should be using absolute imports to get at things inside
the package.

This change may break some of the notebooks, in particular the one
that uses mismatch_comparison.py and comparison_hook.py.  I’ll come
back and fix that later, before merging the branch.