Remove all the unit tests.
Michael says the unit tests are too far out of date to be salvaged,
it would be easier to start from scratch.

If unit tests are ever added again in the future, they should be
placed in a top-level directory named ‘unit_tests’, *not* inside
the pdr_tests package.  The pytest documentation recommends this
organization for new code, because it avoids various problems with
module naming and imports.