Move unit tests to top level and fix test enumeration.
The tests (all two of them) still fail, because MockData is way out of
date with respect to how pdr.Data works anymore, but at least now test
enumeration does not crash.

Having the tests outside the package is recommended by the current
pytest documentation, and like the notebooks, it makes clear how
you are supposed to import stuff from the pdr_tests package to the
unit tests.

(Note: the unit tests we have, appear to be unit tests of pdr itself,
and should maybe be moved to the pdr repository.)