Introduce ‘ix’ entry point in package metadata.
This means that after ‘pip install -e .’ it will be possible to
run ‘ix’ directly as a command line program, instead of having
to do ‘python pdr_tests/ix.py’.

As with the previous patch, several imports within the pdr_tests
package needed to be changed from absolute to relative in order to
accomplish this, and this was tested by manually importing pdr_tests
and all of its submodules from outside the package.  In addition,
‘ix count’ and ‘ix test’ were confirmed to work when run with no
arguments.