Remove unnecessary imports from */selection_rules.py.
19 existing selection_rules.py files were unnecessarily importing
itertools, pathlib, and/or pdr_tests.  Six of these were also
unnecessarily calculating MANIFEST_DIR.  Remove all of that.

In addition to not doing unnecessary work, this change means that
almost all of the places that calculate MANIFEST_DIR based on the
pdr_tests package directory, instead of getting it from the settings,
are gone as well.