Remove unnecessary imports from */selection_rules.py.
16 existing selection_rules.py files were unnecessarily importing
pdr_tests and/or pathlib.  Five of these were also unnecessarily
calculating MANIFEST_DIR.  Remove all of that.

In addition to not doing unnecessary work, this change means that
the name “pdr_tests” is not hardcoded anywhere within pdr_tests,
and almost all of the places that calculate MANIFEST_DIR, instead
of relying on settings.base to do it, are gone as well.