Abolish DatasetDefinition.temp_path.
The only thing this was used for, was to hold partially downloaded
files.  It is better to download into a file with a temporary *name*
in the same *directory* as the final pathname, because this guarantees
that it will be possible to replace any previous version of the file
atomically, with a single call to os.rename().  Also, it means we are
no longer polluting the ix user’s home directory.  And the process
can be completely encapsulated within utilz.ix_utilz._download_chunk,
which means fewer arguments all the way up through verbose_temp_download.