Remove inconsistently used ‘--debug’ option.
The effect of ‘--debug’ was inconsistent across various ix subcommands.
In ‘check’ and ‘test’ it controlled whether we passed ‘debug=True’
to pdr.read and the default was True.  This option is now spelled
‘--pdr-debug’.

In ‘index_directory’ instead it meant “stop on first error and print a
full traceback” and the default was False.  This option is now spelled
‘--stop-on-first-error’.

The new spellings do not have associated short options.

For internal consistency, variables named ‘debug’ whose function is,
ultimately, to control whether we call pdr.read(debug=True) are now
all named ‘pdr_debug’ instead.