Use argparse instead of fire for CLI argument processing.
argparse makes it easier to document shared options, offers more
flexible handling of option arguments, and it’s part of the Python
stdlib.

Shim logic in the new module utilz.cli_utilz derives as much
information as possible from the function signatures of the CLI
entry point functions, so what we have to write in ix_interface
is not much more verbose than it was with fire.   Example:

```
@cli_action(
    dry_run = {
        "short": "d",
        "help": "Do not actually download any labels"
    },
    **COMMON_ARGUMENTS,
)
def index(
    dataset: str,
    product_type: Optional[str] = None,
    *,
    dry_run: bool = False,
):
    """
    Download detached labels and create a comprehensive index for the
    products selected by 'ix pick'.
    """
```

Note that fire is still a dependency of the hostess library and
therefore it still appears in environment*.yml.